[#113076247] Refactor documents edit
Enable editing documents in the document details page
