+++ app/components/Dashboard/DocList.jsx
@@ -2,18 +2,13 @@
   'use strict';
 
   var React = require('react'),
-    DocEdit = require('./DocEdit.jsx'),
-    RoleActions = require('../../actions/RoleActions'),
-    RoleStore = require('../../stores/RoleStore'),
     cardImage = require('../../images/soccer.jpeg');
 
   class DocList extends React.Component {
 
     static propTypes = {
       deleteDoc: React.PropTypes.func,
-      docs:  React.PropTypes.arrayOf(React.PropTypes.object),
-      roles: React.PropTypes.arrayOf(React.PropTypes.object),
-      updateDocs: React.PropTypes.func
+      docs:  React.PropTypes.arrayOf(React.PropTypes.object)
     };
 
     constructor(props) {
@@ -22,18 +17,11 @@
       this.state = {
         docs: this.props.docs,
         deletedDoc: null,
-        roles: null,
         token: localStorage.getItem('user')
       };
-
-      this.handleRolesResult = this.handleRolesResult.bind(this);
     }
 
     componentDidMount() {
-      RoleStore.addChangeListener(this.handleRolesResult);
-
-      // Send a request to check the logged in user
-      RoleActions.getRoles(this.state.token);
       // Activate the materialize tooltips
       setTimeout(function() {
         window.$('.tooltipped').each(function() {
@@ -42,39 +30,10 @@
       }, 1000);
     }
 
-    componentWillReceiveProps(nextProps) {
-      if (nextProps.docs !== this.state.docs) {
-        this.setState({
-          docs: nextProps.docs
-        });
-      }
-    }
-
-    componentWillUnmount() {
-      RoleStore.removeChangeListener(this.handleRolesResult);
-    }
-
-    handleRolesResult() {
-      let roles = RoleStore.getRoles();
-      this.setState({roles: roles});
-    }
-
-    handleDocumentEdit(doc, event) {
-      // Prevent the default action for clicking on a link
-      event.preventDefault();
-      // Get the id of the <a> tag that triggered the modal
-      let id = `#${event.currentTarget.getAttribute('href')}`;
-      // Open the specific modal when the link is clicked
-      window.$(id).openModal();
-    }
-
     render() {
       let renderDoc = function(doc) {
         return (
           <div className="col s12 m6 l4" key={doc._id}>
-            <div id={`edit-modal-${doc._id}`} className="modal">
-              <DocEdit doc={doc} roles={this.state.roles} updateDocs={this.props.updateDocs}/>
-            </div>
             <div className="card">
               <div className="card-image">
                 <img src={cardImage}/>
